+++ composer.json
@@ -35,6 +35,11 @@
             "Ciareis\\Bypass\\": "src"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\" : "tests/"
+        }
+    },
     "extra": {
         "laravel": {
             "providers": [
